# Take Writer/Reader by `&mut` in consensus en/decoding