@dpc 2022-06-03
In debug build (`cargo test`) the differene is around 900KB:

old:
```
> ls -al target/debug/deps/bitcoin-32a2c32b15dd4dc4
-rwxrwxr-x 1 dpc dpc 35997144 Jun  2 23:46 target/debug/deps/bitcoin-32a2c32b15dd4dc4
```
new:
```
> ls -al target/debug/deps/bitcoin-32a2c32b15dd4dc4
-rwxrwxr-x 1 dpc dpc 35097536 Jun  2 23:45 target/debug/deps/bitcoin-32a2c32b15dd4dc4
```