@apoelstra 2022-06-06
The first commit now fails to comipile with
```
error[E0046]: not all trait items implemented, missing: `consensus_decode`
   --> /tmp/tmpzshlvtp4/src/util/psbt/mod.rs:300:1
    |
300 | impl Decodable for PartiallySignedTransaction {
    | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ missing `consensus_decode` in implementation
    |
   ::: /tmp/tmpzshlvtp4/src/consensus/encode.rs:365:5
    |
365 |     fn consensus_decode<R: io::Read>(reader: &mut R) -> Result<Self, Error>;
    |     ------------------------------------------------------------------------ `consensus_decode` from trait

For more information about this error, try `rustc --explain E0046`.
error: could not compile `bitcoin` due to previous error
```