@apoelstra 2022-06-24
I'd still like a third ACK here because it's a fairly big change to the encoding API