@RCasatta 2022-06-28
I tested downstream https://github.com/RCasatta/blocks_iterator/pull/56

I don't fully grasp the change I had to make to this line:

`let len = block_extra.consensus_encode(&mut self.buffer[..]).unwrap();`

where `buffer` is a `Vec<u8>`.

To solve:

```
error[E0277]: the size for values of type `[u8]` cannot be known at compilation time
   --> src/pipe.rs:41:52
    |
41  |             let len = block_extra.consensus_encode(&mut self.buffer[..]).unwrap(); // buffer is big enough, we can unwrap
    |                                   ---------------- ^^^^^^^^^^^^^^^^^^^^ doesn't have a size known at compile-time
    |                                   |
    |                                   required by a bound introduced by this call
    |
    = help: the trait `Sized` is not implemented for `[u8]`
note: required by a bound in `consensus_encode`
   --> /home/casatta/.cargo/git/checkouts/rust-bitcoin-804a7fa56d920f24/a24a3b0/src/consensus/encode.rs:318:25
    |
318 |     fn consensus_encode<W: io::Write>(&self, writer: &mut W) -> Result<usize, io::Error>;
    |                         ^ required by this bound in `consensus_encode`
```
To fix the error I had to pass the `Vec` directly instead of a mutable slice

`let len = block_extra.consensus_encode(&mut self.buffer).unwrap();`