@apoelstra 2022-06-28
@RCasatta the issue is that `consensus_encode` is taking a `&mut W` so if you give it a `&mut [u8]` it will infer that `W` is `[u8]`. Then because we don't have a `Sized?` bound on `W`, it complains because it doesn't know the size of `[u8]`.

Having said this, I'm a little surprised that this code ever worked ... does `&mut [u8]` impl `Write`? How does it know what place to write new data to?