@dpc 2022-06-28
> Having said this, I'm a little surprised that this code ever worked ... does `&mut [u8]` impl `Write`? How does it know what place to write new data to?

https://doc.rust-lang.org/std/io/trait.Write.html#impl-Write-6