@dpc 2022-06-28
On `bitcoin_iterators` examples we're seeing 50% of size improvement?! (seems too good to be true)? Haven't got a chance to run the perf.



> @RCasatta the issue is that `consensus_encode` is taking a `&mut W` so if you give it a `&mut [u8]` it will infer that `W` is `[u8]`. Then because we don't have a `Sized?` bound on `W`, it complains because it doesn't know the size of `[u8]`.

Should I change to `<W: io::Write + ?Sized>` then?