@dpc 2022-06-28
AFAIR, `: Sized?` makes the `T` just harder to store  - since it doesn't have a statically known size (though it is possible to `Box` it or something). That's why it's not the default, because 99% of the time people want to store their `T` somewhere. In this case we only ever use the `writer` as variable we pass around, plus we take a reference to it anyway, so should be OK.

The only downside I see (other than me having to go and do more work) is the users having to do more work by adding this `: Sized?`. But I guess we should do it to unlock maximum potential.