@RCasatta 2022-06-29
ACK 1fea098dfb916d64f2ecbfb9a13b8d264020bc1e tested in downstream lib, space saving in compiled code confirmed

Thanks a lot for adding `?Sized` I think it's very useful not for the example I previously mentioned but when you want to write on mutable slice not from the beginning, see for example https://github.com/RCasatta/blocks_iterator/blob/27d3cd2d75a36167aed3de514136383df73cbf0b/src/utxo/db.rs#L51

The only thing to mention is the slightly surprising syntax to write in a mutable slice, requiring now (IIUC) a double `&mut`: https://github.com/RCasatta/blocks_iterator/blob/e8ef94e3ef8f3a8723a80aa9ca980a3fee334520/src/pipe.rs#L42